+++ src/theVoice/TheVoice.tsx
@@ -1,21 +1,17 @@
 import React, { ReactNode } from 'react'
-import cc from 'classcat'
 
 import { StyledTheVoice } from './TheVoice.style'
 
 export type TheVoiceProps = Readonly<{
+  children: ReactNode
   id?: string
   className?: string
-  children: ReactNode
   isInverted?: boolean
+  as?: keyof JSX.IntrinsicElements | React.ComponentType<any> // from styled-components types
 }>
 
-export const TheVoice = ({ id, className, children, isInverted }: TheVoiceProps) => (
-  <StyledTheVoice
-    id={id}
-    className={cc([className, { 'kirk-thevoice--inverted': isInverted }])}
-    headingLevel="1"
-  >
+export const TheVoice = ({ id, className, children, isInverted, as = 'h1' }: TheVoiceProps) => (
+  <StyledTheVoice id={id} as={as} className={className} isInverted={isInverted}>
     {children}
   </StyledTheVoice>
 )