+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+**[UPDATE]** Allow polymorphism with `as` prop for `TheVoice`
 - [...]
 
 # v55.0.0 (18/03/2021)